test(NODE-7623): extract section 12 explicit encryption to own file
Move the '12. Explicit Encryption' prose tests out of the monolithic
client_side_encryption.prose.test.ts into their own file following the
pattern of other prose test files. Pure code move with no behavior
changes; the eeMetadata/eeMetadataPre90 definitions used only by this
section move with it, and getKmsProviders is imported from the
monolithic file which still exports it.
